Add department dropdown to form
This commit adds the department dropdown
back to the "About Me" portion of the form.

Right now they are both on the "About Me" tab,
but the workflow diagrams have the department
dropdown on the next tab.

This changes the name of the `formSchema` object
to `formStore` to better reflect what that object
is.  The object is a store from the Flux pattern:
https://vuejs.org/v2/guide/state-management.html#Simple-State-Management-from-Scratch

That means that any changes to the state of the `formStore` object
should be done using methods on the `formStore` object itself. That
way the object can be used in many different components, with all
of the changes happening in a known location.